requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.18
requests==2.31.0
beautifulsoup4==4.12.2

# Data analysis and ML
scikit-learn==1.3.2
ta==0.10.2
textblob==0.17.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Visualization
plotly==5.17.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3 